www: fix "500 internal server error"
A dangling symlink in /mutable/TeslaCam would cause the Recordings
tab in the web interface to fail with a "500 internal server error"
because nginx doesn't like it when stat() fails..
Avoid this by having cttseraser lstat() the path when stat() fails.
If lstat() succeeds where stat() didn't, return the stat info for the
link itself. Nginx will then try to get info for the link itself, which
will fail because cttseraser doesn't implement that functiom. This in
turn will make nginx ignore the link altogether and not even show it in
the file list.